Capitalise "GL" and "GLU" to fix build errors:
It previously worked with these being lowercase. Since then, I upgraded my system from Debian 11 to 12.
However, I'm not sure if the change in case is a result of my Debian upgrade, or if lowercase was always wrong and my system wrongly accepted it before.
I've updated README.md as well, to confirm that Debian 12 works.